added look up controller
