+++ src/types/page.urls.ts
@@ -198,11 +198,10 @@ export const UPDATE_DIRECTOR_SUBMITTED_PATH_END = UPDATE_PATH + "director-submit
 export const UPDATE_DIRECTOR_SUBMITTED = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_SUBMITTED_PATH_END;
 export const UPDATE_DIRECTOR_SUBMITTED_PATH = OFFICER_FILING + UPDATE_DIRECTOR_SUBMITTED;
 
-export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END = UPDATE_PATH + "director-correspondence-address-manual";
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH_END = UPDATE_PATH + "find-director-correspondence-address";
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH_END;
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH;
+
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END = UPDATE_PATH + "enter-director-correspondence-address";
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END;
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL;
-
-export const UPDATE_DIRECTOR_CONFIRM_CORRESPONDENCE_ADDRESS_PATH_END = UPDATE_PATH + "confirm-director-correspondence-address";
-export const UPDATE_DIRECTOR_CONFIRM_CORRESPONDENCE_ADDRESS = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CONFIRM_CORRESPONDENCE_ADDRESS_PATH_END;
-export const UPDATE_DIRECTOR_CONFIRM_CORRESPONDENCE_ADDRESS_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CONFIRM_CORRESPONDENCE_ADDRESS;
-
