+++ views/includes/page-templates/director.correspondence.address.search.html
@@ -1,0 +1,45 @@
+{% extends "layout.html" %}
+
+{% set title = "Find the director's correspondence address" %}
+
+{% block pageTitle %}
+  {% include "includes/page-title.html" %}
+{% endblock %}
+
+{% block backLink %}
+  {% set backLinkText = "Go back to 'What is the director's correspondence address'" %}
+  {% include "includes/back-link.html" %}
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-three-quarters">
+
+      {% include "includes/list/errors.html" %}
+      
+      <form action="" method="post">
+        <span class="govuk-caption-l">{{directorName}}</span>
+        <h1 class="govuk-heading-l">Find the director's correspondence address</h1>
+        <!-- This is the search page for the correspondence address -->
+        {% from "govuk/components/input/macro.njk" import govukInput %}
+        <div id="typeahead-hint" class="govuk-hint">
+          We'll use this address to send Companies House mail to the director. You can use a business or home address and it does not have to be in the UK.
+        </div>
+        {% include "includes/inputs/fields/postcode-input.html" %}
+        {% include "includes/inputs/fields/postcode-premises-input.html" %}
+
+        {{ govukButton({
+        attributes: {
+        "data-event-id": "continue-button"
+        },
+        text: "Find address"
+        }) }}
+
+        <div class="govuk-body">
+          <a href="{{enterAddressManuallyUrl}}" class="govuk-link" data-event-id="director-correspondence-address-manual-link">Enter address manually</a>
+        </div>
+
+      </form>
+    </div>
+  </div>
+{% endblock %}
